refactor(OrderTable): Order와 OrderTable 패키지 분리
- Order에 묻는 로직은 Order에서 구현
  - 대신 Interface를 OrderTable에서 제공하고, OrderTable은 해당 구현된 빈을 주입받는다.
