+++ src/main/java/kitchenpos/table/domain/OrderTableValidatable.java
@@ -1,0 +1,5 @@
+package kitchenpos.table.domain;
+
+public interface OrderTableValidatable {
+    void validateHasProgressOrder(OrderTable orderTable);
+}
