+++ src/main/java/kitchenpos/table/dto/OrderTableResponse.java
@@ -1,7 +1,6 @@
-package kitchenpos.order.dto;
+package kitchenpos.table.dto;
 
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.table.domain.OrderTable;
 
 import java.util.List;
 import java.util.Objects;
