+++ src/main/java/kitchenpos/table/exception/OrderTableNotFoundException.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.exception;
+package kitchenpos.table.exception;
 
 import kitchenpos.common.exception.NotFoundException;
