+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -9,9 +9,9 @@ import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.menu.exception.MenuNotFoundException;
 import kitchenpos.order.exception.OrderNotFoundException;
-import kitchenpos.order.exception.OrderTableNotFoundException;
+import kitchenpos.table.exception.OrderTableNotFoundException;
 import kitchenpos.product.domain.Product;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import org.assertj.core.api.ThrowableAssert;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
