+++ src/test/java/kitchenpos/tablegroup/acceptance/TableGroupAcceptanceTest.java
@@ -4,8 +4,8 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.AcceptanceTest;
 import kitchenpos.order.acceptance.OrderAcceptanceTest;
-import kitchenpos.order.acceptance.TableAcceptanceTest;
-import kitchenpos.order.dto.OrderTableResponse;
+import kitchenpos.table.acceptance.TableAcceptanceTest;
+import kitchenpos.table.dto.OrderTableResponse;
 import kitchenpos.tablegroup.dto.OrderTableRequest;
 import kitchenpos.tablegroup.dto.TableGroupRequest;
 import kitchenpos.tablegroup.dto.TableGroupResponse;
